fix(adc): drop redundant u32 cast in apb_saradc oneshot test
`(1650u32 * 4095) / 3300` is already u32, so the trailing `as u32` is a
no-op and clippy's `unnecessary_cast` rejects it under pr-gate's
`-D warnings`. The `as u16` three lines up is a real narrowing cast and
stays.

Only `clippy --all-targets` sees this — the expression is inside a
`#[cfg(test)]` block, so `cargo test --lib set_channel_input_overrides`
passes with the cast in place.

+++ crates/core/src/peripherals/esp32c3/apb_saradc.rs
@@ -408,7 +408,7 @@ mod tests {
         assert_eq!(a.channel_input_count(1), ((1650u32 * 4095) / 3300) as u16);
         a.write_u32(ONETIME_SAMPLE, sar1_oneshot(1)).unwrap();
         let data = a.read_u32(SAR1DATA_STATUS).unwrap() & 0x0FFF;
-        let expected = ((1650u32 * 4095) / 3300) as u32;
+        let expected = (1650u32 * 4095) / 3300;
         assert_eq!(data, expected, "oneshot must return the injected count");
         // Un-injected channel still uses the ramp.
         a.write_u32(ONETIME_SAMPLE, sar1_oneshot(2)).unwrap();
